backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
anthropic==0.34.2
python-dotenv==1.0.1
